requirements.txt
numpy<2.0
requests==2.32.4
python-dotenv==1.1.1
ollama~=0.6.0
mysql-connector-python~=9.3.0
chromadb~=1.3.0